+++ lib.commonjs/_tests/test-utils-maths.js
@@ -170,4 +170,25 @@ describe("Tests Bad Math Values", function () {
         });
     });
 });
+describe("Tests Twos Compliemnts Functions", function () {
+    const tests = [
+        { width: 8, value: 0, twos: 0 },
+        { width: 8, value: 1, twos: 1 },
+        { width: 8, value: -1, twos: 0xff },
+        { width: 8, value: 127, twos: 127 },
+        { width: 8, value: -128, twos: 0x80 },
+    ];
+    for (const { twos, width, value } of tests) {
+        it(`computes twos compliment values: ${value}[${width} bits]`, function () {
+            const result = (0, index_js_1.toTwos)(value, width);
+            assert_1.default.equal(result, twos);
+        });
+    }
+    for (const { twos, width, value } of tests) {
+        it(`computes values from twos compliment: ${value}[${width} bits]`, function () {
+            const result = (0, index_js_1.fromTwos)(twos, width);
+            assert_1.default.equal(result, value);
+        });
+    }
+});
 //# sourceMappingURL=test-utils-maths.js.map