+++ lib.commonjs/_tests/test-wallet-hd.d.ts
@@ -1,6 +1,1 @@
-declare global {
-    class TextDecoder {
-        decode(data: Uint8Array): string;
-    }
-}
 export {};