+++ lib.esm/_admin/update-version.js
@@ -60,15 +60,20 @@ function writeVersion(version) {
         }
         pkgInfo.gitHead = gitHead;
         // Save the package.json
-        const check = { "require": 1, "import": 1, "types": 1 };
+        const check = { "default": 1, "require": 1, "import": 1, "types": 1 };
         saveJson(pkgPath, pkgInfo, (path, a, b) => {
-            if (path.startsWith("./") && check[a] && check[b]) {
-                if (a === "types") {
+            if ((path.startsWith("./") || path === ".") && check[a] && check[b]) {
+                const cmp = a.localeCompare(b);
+                if (cmp === 0) {
+                    return cmp;
+                }
+                if (a === "types" || b === "default") {
                     return -1;
                 }
-                if (b === "types") {
+                if (b === "types" || a === "default") {
                     return 1;
                 }
+                return cmp;
             }
             return a.localeCompare(b);
         });