+++ lib.esm/_tests/test-utils-maths.js
@@ -1,5 +1,5 @@
 import assert from "assert";
-import { isError, getBigInt, getNumber, toBeArray, toBeHex, toQuantity, } from "../index.js";
+import { isError, fromTwos, toTwos, getBigInt, getNumber, toBeArray, toBeHex, toQuantity, } from "../index.js";
 describe("Tests Quantity Functions", function () {
     const quantities = [
         {
@@ -164,5 +164,26 @@ describe("Tests Bad Math Values", function () {
                 e.message.startsWith("unsigned value cannot be negative"));
         });
     });
+});
+describe("Tests Twos Compliemnts Functions", function () {
+    const tests = [
+        { width: 8, value: 0, twos: 0 },
+        { width: 8, value: 1, twos: 1 },
+        { width: 8, value: -1, twos: 0xff },
+        { width: 8, value: 127, twos: 127 },
+        { width: 8, value: -128, twos: 0x80 },
+    ];
+    for (const { twos, width, value } of tests) {
+        it(`computes twos compliment values: ${value}[${width} bits]`, function () {
+            const result = toTwos(value, width);
+            assert.equal(result, twos);
+        });
+    }
+    for (const { twos, width, value } of tests) {
+        it(`computes values from twos compliment: ${value}[${width} bits]`, function () {
+            const result = fromTwos(twos, width);
+            assert.equal(result, value);
+        });
+    }
 });
 //# sourceMappingURL=test-utils-maths.js.map