+++ lib.esm/_tests/test-wallet-hd.js
@@ -2,6 +2,13 @@ import assert from "assert";
 import { wordlists } from "../wordlists/wordlists.js";
 import { loadTests } from "./utils.js";
 import { HDNodeWallet, HDNodeVoidWallet, Mnemonic } from "../index.js";
+/*
+declare global {
+    class TextDecoder {
+        decode(data: Uint8Array): string;
+    }
+}
+*/
 const decoder = new TextDecoder();
 function fromHex(hex) {
     const data = Buffer.from(hex.substring(2), "hex");